Add shstyle(9) to TODO list
Something I've wanted to create for a long time, this would be the
style(9) equivalent for sh(1) scripts to help make sure that any
scripts created for the project are compatible when necessary and
are as performant as possible. This will also be where I lay out
information about standardized layout(s).